Update rkVerifyBackups.php
Updates : support Windows OS !

+++ rkVerifyBackups.php
@@ -1,27 +1,37 @@
 <?php
 	// Include section
 	include_once 'rkFramework.php';
+	include_once 'rkCredentials.php';
 
-	// Initialisation section
-	$clusterConnect=array(
-		"username" => "username",
-		"password" => "password",
-		"ip" => "<cluster ip>"
-	);
-
-	$objectID="NutanixVirtualMachine:::86be7f0c-2b98-4e03-93d9-57f06e419eb7-vm-c1dd1716-7d54-4121-9434-1a12ea0fb37e";
-	$snapID="26eba277-8eee-4c7e-9c12-9671a82367a6";
 	$tempFolder=sys_get_temp_dir();
-		
+
+	//$object="Centos-vm1";
+	$object="Win10-vm1";
+	
+	// Detects what OS the script is running on. Either Windows or Linux family	
+	$os=PHP_OS;
+
+	// For Windows only
+	// GNU Wget for Windows is required for this script to work -> https://eternallybored.org/misc/wget/
+	
+	$wgetPath="C:\\";
+
+	if($os=="WINNT" && !file_exists($wgetPath."wget.exe"))
+	{
+		print(rkColorRed("ERROR, WGET.EXE not found in ".$wgetPath.".\n\n"));
+		exit();
+	} 	
+	
 	// ===========================================
 	// Main entry point
 	// ===========================================
 
-	print("\n===============================\n");
+	print("\n\n\n\n===============================\n");
 	print(" Rubrik Backup Integrity Check\n");
-	print("=======================v 1.0===\n\n");
+	print("=======================v 1.1===\n\n");
 
-	print("Target Object : ".rkColorOutput($objectID)."\n");
+	$objectID=rkObjectNametoID($clusterConnect,$object);
+	print("Target Object : ".rkColorOutput($object)." (ID : ".rkColorOutput($objectID).")\n");
 
 	// If snapshot ID has been specify, use it, either the last available snap for object will be selected by default
 	if (isset($snapID))
@@ -55,10 +65,22 @@
 		
 	print(rkColorOutput("\nDone!      \n"));
 
-	system("wget -q --no-check-certificate -O ".$tempFolder."/".$eventID." ".$result->links[1]->href." > /dev/null");
-	$content=file_get_contents($tempFolder."/".$eventID);
-	system("rm ".$tempFolder."/".$eventID);
-
+	if($os=="WINNT")
+	{
+		// Windows does not support ":" as part of the file name, change with "_"
+		$tempFileName=str_replace(":","_",$eventID);
+		system($wgetPath."wget.exe -q --no-check-certificate -O ".$tempFolder."\\".$tempFileName." ".$result->links[1]->href." > NULL");
+		$content=file_get_contents($tempFolder."\\".$tempFileName);
+		system("del ".$tempFolder."\\".$tempFileName);
+	}	
+	else
+	{
+		// If not Windows, we are in UNIX Family so we are using /
+		system("wget -q --no-check-certificate -O ".$tempFolder."/".$eventID." ".$result->links[1]->href." > /dev/null");
+		$content=file_get_contents($tempFolder."/".$eventID);
+		system("rm ".$tempFolder."/".$eventID);
+	}	
+	
 	// Compute time taken to verify snapshot
 	$startTime=$result->startTime;
 	$endTime=$result->endTime;
@@ -68,7 +90,7 @@
 	$endTS=rkGetTimeStamp($endTime);
 	$duration=$endTS-$startTS;
 	$duration=date($dateFormat,$duration);	
-	
+
 	$tmp=explode(",",$content);
 	$integrityStatus=str_replace("\"","",$tmp[16]);
 	$objectName=str_replace("\"","",$tmp[11]);
